refactor proposal narrative routes
